+++ generators/csharp/codegen/src/ast/core/Writer.ts
@@ -25,7 +25,7 @@ export declare namespace Writer {
 
 export class Writer {
     /* The contents being written */
-    private buffer = "";
+    public buffer = "";
     /* Indentation level (multiple of 4) */
     private indentLevel = 0;
     /* Whether anything has been written to the buffer */
@@ -207,10 +207,6 @@ export class Writer {
     public importsToString(): string | undefined {
         const imports = this.stringifyImports();
         return imports.length > 0 ? imports : undefined;
-    }
-
-    public getBuffer(): string {
-        return this.buffer;
     }
 
     public isReadOnlyMemoryType(type: string): boolean {